requirements.txt
aiohttp>=3.8.0
asyncpg>=0.28.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
coloredlogs>=15.0.0